Use new WCSR framework unit testcase class to house helper functions
